added mongo pipeline

+++ scripts/mapreduce/Q1Reducer.java
@@ -8,14 +8,14 @@ public class Q1Reducer extends Reducer<Text, Text, Text, Text> {
     public void reduce(Text key, Iterable<Text> values, Context context)
             throws IOException, InterruptedException {
 
-        int requestCount = 0;
-        int totalBytes = 0;
+        long requestCount = 0;
+        long totalBytes = 0;
 
         for (Text val : values) {
             String[] parts = val.toString().split("_");
 
-            requestCount += Integer.parseInt(parts[0]);
-            totalBytes += Integer.parseInt(parts[1]);
+            requestCount += Long.parseLong(parts[0]);
+            totalBytes += Long.parseLong(parts[1]);
         }
 
         // Output: date_status    requestCount_totalBytes
